LayoutTests/fast/canvas/canvas-createPattern-video-modify-expected.txt

Tests that modifying the video used in createPattern does not affect the pattern rendered.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS buffer[0] is within 2 of 200
PASS buffer[1] is within 2 of 200
PASS buffer[2] is within 2 of 0
PASS buffer[0] is within 2 of 200
PASS buffer[1] is within 2 of 200
PASS buffer[2] is within 2 of 0
PASS buffer[0] is within 2 of 200
PASS buffer[1] is within 2 of 200
PASS buffer[2] is within 2 of 0
PASS buffer[0] is within 2 of 200
PASS buffer[1] is within 2 of 200
PASS buffer[2] is within 2 of 0
PASS successfullyParsed is true

TEST COMPLETE
